Separated components into Sphere and ThreeDObj s
