+++ public/styles.css
@@ -1,3 +1,3 @@
 main {
-  background-color: rgb(0, 0, 0);
+  background-color: #ffffff;
 }
